Made basic memory manager to dish out new pages.
The next step is to distribute stacks for threads and add a guard page.
